+++ src/main/java/com/the_pathfinders/InitialController.java
@@ -16,8 +16,6 @@ import javafx.util.Duration;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-import javax.swing.Action;
-
 public class InitialController implements Initializable {
 
     @FXML private Pane root;